Comprobado que funciona el metodo para Get/Set desde el SimManager

+++ Assets/Scripts/Netcode/Simulation/BulletSimulationState.cs
@@ -8,9 +8,9 @@ public struct BulletSimulationState : ISimulationState
     public Vector2 Position { private set; get; }
     public float Rotation { private set; get; }
     public Vector2 Velocity { private set; get; }
-    public int LifeTime { private set; get; }
+    public float LifeTime { private set; get; } // In seconds
 
-    public BulletSimulationState(Vector2 position, float rotation, Vector2 velocity, int lifeTime)
+    public BulletSimulationState(Vector2 position, float rotation, Vector2 velocity, float lifeTime)
     {
         this.Position = position;
         this.Rotation = rotation;
